Cambios de lineas y arreglos
Se borraron declaraciones que no se usaban, se modificaron algunas funciones para que acepten un digito mas alto

+++ Camping Sol Luna/include/Funciones.h
@@ -3,5 +3,5 @@
 #include <iostream>
 #include <limits>
 using namespace std;
-int IngresoNumero();
+long long int IngresoNumero();
 #endif // FUNCIONES_H_INCLUDED
