+++ Camping Sol Luna/src/Funciones.cpp
@@ -2,15 +2,17 @@
 #include <limits>
 #include <string>
 #include <sstream>
+#include <cctype>
+#include <cstring>
 using namespace std;
 
 #include "Funciones.h"
 
 ///Alan
 ///Valida que el ingreso sea un numero y que sea dntro de los rngos posibles
-int IngresoNumero()
+long long int IngresoNumero()
 {
-    int numeros;
+    long long int numeros;
     while (!(cin >> numeros)) {
         cin.clear();
         cin.ignore(numeric_limits<streamsize>::max(), '\n');
@@ -20,5 +22,4 @@ int IngresoNumero()
 
     return numeros;
 }
-
